half of the pages are refatored to new design and functionality.
