+++ models/Logout.php
