+++ views/admin/BikeSpecificationsView.php
