+++ views/admin/BrakeSystemView.php
@@ -59,16 +59,20 @@
     <?php endif  ?>
 
     <!--    the form for creating and updating drive_type starts here-->
-    <div class="modal fade" id="exampleModalCenter" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
+    <div class="modal fade <?php echo isset($_POST["update"]) ? 'show' : ' ' ?>" id="exampleModalCenter" tabindex="-1" role="dialog"
+         aria-labelledby="exampleModalCenterTitle" aria-hidden="true"
+        <?php echo isset($_POST["update"]) ? 'style = "display : block"' : 'style = "display : none"'?>>
         <div class="modal-dialog modal-dialog" role="document">
             <div class="modal-content">
                 <div class="modal-header">
                     <h5 class="modal-title" id="exampleModalLongTitle">
                         <?php echo !$br->getUpdate() ? "Create new drive type" : "Update product " . $br->getVal()['name'] ?>
                     </h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
+                    <form action="" method="post">
+                        <button type="submit" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                    </form>
                 </div>
                 <div class="modal-body">
                     <form class="col-12" action="" method="post" id="form">
@@ -97,7 +101,8 @@
                             <input type="text" hidden name = "braking_systemID" value = "<?php echo $br->getVal()['braking_systemID'] ?>">
                         <?php endif; ?>
                         <div class="form-group col-12 mt-2">
-                            <input type="submit" class="btn <?php echo !$br->getUpdate() ? 'btn-primary' : 'btn-info' ?>" name="<?php echo !$br->getUpdate() ? 'submit-new' : 'submit-update' ?>" value="<?php echo !$br->getUpdate() ? 'Create new' : 'update' ?>">
+                            <input type="submit" class="btn <?php echo !$br->getUpdate() ? 'btn-primary' : 'btn-info' ?>" name="<?php echo !$br->getUpdate() ? 'submit-new' : 'submit-update' ?>" value="<?php echo !$br->getUpdate() ? 'Create new' : 'update' ?>"
+                            <input type="submit" class="btn btn-secondary" value = "Cancel">
                         </div>
                     </form>
                 </div>
