+++ views/admin/ImageView.php
