--- web-shop/includes/adminNavigation.php
@@ -1,19 +1,0 @@
-    <div class="tab_nav_wrapper" style="width: 20vw; height:100vh; background-color:whitesmoke; border-right:rgba(0,0,0,0.10) solid 0.5px;">
-        <div class="tab_nav_container" style="padding:1rem;width:100%; height:100%; display:flex; flex-direction:column; justify-content:flex-start; align-items:center;">
-                <div style="display: flex; justify-content:flex-start; align-items:center; height:auto; width:100%;">
-                    <img style="width: 4rem; height:4rem;" src="/bike-shop-solution/web-shop/public/owl-logo.svg" alt="">
-                </div>
-                <div style="width: 100%; margin-top:8rem;">
-                    <div style="width: 100% !important;" class="nav flex-column nav-pills" id="v-pills-tab" role="tablist" aria-orientation="vertical">
-                        <a href="/bike-shop-solution/web-shop/admin-category" style="font-size:1.2rem; width: 100% !important; color:black; margin-bottom:0.5rem;">Categories</a>
-                        <a href="/bike-shop-solution/web-shop/admin-drive-type" style="font-size:1.2rem; width: 100% !important; color:black; margin-bottom:0.5rem;">Drive Types</a>
-                        <a href="/bike-shop-solution/web-shop/admin-brake-system" style="font-size:1.2rem; width: 100% !important; color:black; margin-bottom:0.5rem;">Break Systems</a>
-                        <a href="/bike-shop-solution/web-shop/admin-wheel-size" style="font-size:1.2rem; width: 100% !important; color:black; margin-bottom:0.5rem;">Wheel Sizes</a>
-                        <a href="/bike-shop-solution/web-shop/admin-company-details" style="font-size:1.2rem; width: 100% !important; color:black; margin-bottom:0.5rem;">Owl Bikeshop Details</a>
-                    </div>
-                </div>
-            <div style="height:7vh; width: 20vw; position:absolute; bottom:0; left:0; display:flex; align-items:center; padding:1rem; justify-content:flex-start; background-color:rgba(0,0,0,0.15);">
-                <h4 style="color:black; opacity:0.7; font-size:0.8rem; margin:0; padding:0;">Logged in user: Radek or Raul</h4>
-            </div>
-        </div>
-    </div>
